requirements.txt
pypresence
paho-mqtt
icecream
